+++ CHANGELOG.md
@@ -18,6 +18,7 @@ branch is the latest unreleased version.
 	- types.some is now [Optional.some](https://github.com/Anaminus/rbxmk/blob/imperative/doc/types.md#user-content-optionalsome).
 
 **Highlights:**
+- Add `base64` format for encoding and decoding strings in Base64.
 - Add `plugin` dump format, which enables custom dump formats.
 	- Runs a Lua script in a limited environment.
 	- Receives a table that describes the Lua environment, and a function to